added download links for data extraction, needs testing
