+++ arc_application/views/audit_log.py
@@ -26,7 +26,7 @@ def audit_log_dispatcher(request):
     elif request.GET['app_type'] == 'Nanny':
         return HttpResponseRedirect(reverse('nanny-auditlog') + '?id=' + request.GET['id'])
         # return NannyAuditLog.as_view()(request)
-    elif request.GET['app_type'] == 'Association':
+    elif request.GET['app_type'] == 'New Association':
         return HttpResponseRedirect(reverse('hm-auditlog') + '?id=' + request.GET['id'])
     else:
         raise ValueError('The "app_type" request.GET QueryDict does not equal either "Childminder" nor "Nanny" nor "'
