+++ arc_service/urls.py
@@ -36,6 +36,7 @@ from arc_application.views.contact_centre.nanny_change_details import NannyUpdat
 from arc_application.views.search_router import SearchRouter
 from arc_application.views import upload_capita_dbs
 from arc_application.views.applications_summary import ApplicationsSummaryView
+from arc_application.views.daily_reporting import ApplicationsInQueueView, ApplicationsReturnedView, ApplicationsProcessedView, ApplicationsAssignedView
 from arc_application.views.adult_update_views.adult_update_view import new_adults_summary
 from arc_application.views.adult_update_views.adult_previous_registration import adult_previous_registration_view
 from arc_application.views.adult_update_views.adult_update_summary import arc_summary as adult_arc_summary
@@ -72,6 +73,10 @@ urlpatterns = [
     url(r'^summary/', ARCUserSummaryView.as_view(), name='summary'),
     url(r'^upload-capita-dbs/$', upload_capita_dbs, name='Upload-Capita-DBS'),
     url(r'^applications-summary/$', ApplicationsSummaryView.as_view(), name='applications-summary'),
+    url(r'^applications-in-queue/$', ApplicationsInQueueView.as_view(), name='applications-in-queue'),
+    url(r'^applications-returned/$', ApplicationsReturnedView.as_view(), name='applications-returned'),
+    url(r'^applications-processed/$', ApplicationsProcessedView.as_view(), name='applications-processed'),
+    url(r'^applications-assigned/$', ApplicationsAssignedView.as_view(), name='applications-assigned'),
 
     # childminder application review
     url(r'^review/$', task_list, name='task_list'),
